pkg/prometheus: Ignore PartialResponseStrategy for Prometheus
partial_response_strategy field is only applicable for Thanos
so if a user sets this by mistake in PrometheusRule object
prometheus errors this field is not found in rule format.

This commit is to fix that and allow this field only for Thanos

Fixes #5124
