+++ pkg/prometheus/rules.go
@@ -22,19 +22,17 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/ghodss/yaml"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/pkg/errors"
 	"github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring"
 	"github.com/prometheus-operator/prometheus-operator/pkg/k8sutil"
-	"github.com/prometheus/prometheus/model/rulefmt"
+	"github.com/prometheus-operator/prometheus-operator/pkg/operator"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	namespacelabeler "github.com/prometheus-operator/prometheus-operator/pkg/namespace-labeler"
-	thanostypes "github.com/thanos-io/thanos/pkg/store/storepb"
 )
 
 const labelPrometheusName = "prometheus-name"
@@ -205,7 +203,7 @@ func (c *Operator) selectRules(p *monitoringv1.Prometheus, namespaces []string) 
 				return
 			}
 
-			content, err := GenerateContent(promRule.Spec, c.logger)
+			content, err := generateRulesConfiguration(promRule.Spec, c.logger)
 			if err != nil {
 				marshalErr = err
 				return
@@ -331,45 +329,11 @@ func prometheusRuleConfigMapName(prometheusName string) string {
 	return "prometheus-" + prometheusName + "-rulefiles"
 }
 
-// GenerateContent takes a PrometheusRuleSpec and generates the rule content
-func GenerateContent(promRule monitoringv1.PrometheusRuleSpec, logger log.Logger) (string, error) {
-	content, err := yaml.Marshal(promRule)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to marshal content")
-	}
-	errs := ValidateRule(promRule)
-	if len(errs) != 0 {
-		const m = "Invalid rule"
-		level.Debug(logger).Log("msg", m, "content", content)
-		for _, err := range errs {
-			level.Info(logger).Log("msg", m, "err", err)
-		}
-		return "", errors.New(m)
-	}
-	return string(content), nil
-}
-
-// ValidateRule takes PrometheusRuleSpec and validates it using the upstream prometheus rule validator
-func ValidateRule(promRule monitoringv1.PrometheusRuleSpec) []error {
-	for i, group := range promRule.Groups {
-		if group.PartialResponseStrategy == "" {
-			continue
-		}
-		// TODO(slashpai): Remove this validation after v0.65 since this is handled at CRD level
-		if _, ok := thanostypes.PartialResponseStrategy_value[strings.ToUpper(group.PartialResponseStrategy)]; !ok {
-			return []error{
-				fmt.Errorf("invalid partial_response_strategy %s value", group.PartialResponseStrategy),
-			}
-		}
-
-		// reset this as the upstream prometheus rule validator
-		// is not aware of the partial_response_strategy field
+func generateRulesConfiguration(promRule monitoringv1.PrometheusRuleSpec, logger log.Logger) (string, error) {
+	// Unset partialResponseStrategy field.
+	for i := range promRule.Groups {
 		promRule.Groups[i].PartialResponseStrategy = ""
 	}
-	content, err := yaml.Marshal(promRule)
-	if err != nil {
-		return []error{errors.Wrap(err, "failed to marshal content")}
-	}
-	_, errs := rulefmt.Parse(content)
-	return errs
+
+	return operator.GenerateRulesConfiguration(promRule, logger)
 }
